:recycle: refactor: ChallengeCourseProjection 패키지 이동
