+++ src/main/java/com/saisai/domain/challenge/service/ChallengeService.java
@@ -4,7 +4,7 @@ import com.saisai.config.jwt.AuthUserDetails;
 import com.saisai.domain.challenge.dto.response.PopularChallengeCourseRes;
 import com.saisai.domain.challenge.repository.ChallengeRepository;
 import com.saisai.domain.common.aws.s3.ImageUtil;
-import com.saisai.domain.course.dto.projection.ChallengeCourseProjection;
+import com.saisai.domain.challenge.dto.projection.ChallengeCourseProjection;
 import com.saisai.domain.reward.dto.projection.RewardEventProjection;
 import com.saisai.domain.reward.util.RewardUtils;
 import java.util.Collections;
