+++ src/main/java/com/saisai/domain/course/repository/CourseBookmarkRepositoryCustom.java
@@ -1,7 +1,7 @@
 package com.saisai.domain.course.repository;
 
 import com.saisai.domain.course.constant.CourseSortOption;
-import com.saisai.domain.course.dto.projection.ChallengeCourseProjection;
+import com.saisai.domain.challenge.dto.projection.ChallengeCourseProjection;
 import com.saisai.domain.course.dto.projection.GeneralCourseProjection;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
