+++ src/main/java/com/saisai/domain/course/repository/CourseBookmarkRepositoryImpl.java
@@ -9,11 +9,11 @@ import static com.saisai.domain.ride.entity.QRide.ride;
 import com.querydsl.core.types.dsl.Expressions;
 import com.querydsl.jpa.impl.JPAQuery;
 import com.querydsl.jpa.impl.JPAQueryFactory;
+import com.saisai.domain.challenge.dto.projection.ChallengeCourseProjection;
+import com.saisai.domain.challenge.dto.projection.QChallengeCourseProjection;
 import com.saisai.domain.challenge.entity.ChallengeStatus;
 import com.saisai.domain.course.constant.CourseSortOption;
-import com.saisai.domain.course.dto.projection.ChallengeCourseProjection;
 import com.saisai.domain.course.dto.projection.GeneralCourseProjection;
-import com.saisai.domain.course.dto.projection.QChallengeCourseProjection;
 import com.saisai.domain.course.dto.projection.QGeneralCourseProjection;
 import com.saisai.domain.reward.dto.projection.QRewardEventProjection;
 import com.saisai.domain.reward.entity.EventStatus;
